refactor: validate ast
