+++ src/plugin/arguments/src/commonMain/kotlin/community/flock/wirespec/plugin/Wirespec.kt
@@ -1,21 +1,13 @@
 package community.flock.wirespec.plugin
 
-import arrow.core.getOrElse
 import community.flock.wirespec.compiler.core.CompilationContext
 import community.flock.wirespec.compiler.core.ModuleContent
 import community.flock.wirespec.compiler.core.compile
 import community.flock.wirespec.compiler.core.exceptions.WirespecException
-import community.flock.wirespec.compiler.core.parse.ParseOptions
-import community.flock.wirespec.compiler.core.parse.Parser.validate
 import community.flock.wirespec.converter.avro.AvroParser
 import community.flock.wirespec.converter.common.Parser
 import community.flock.wirespec.openapi.v2.OpenAPIV2Parser
 import community.flock.wirespec.openapi.v3.OpenAPIV3Parser
-
-fun ConverterArguments.parseOptions() = ParseOptions(
-    strict = strict,
-    allowUnions = true,
-)
 
 fun compile(arguments: CompilerArguments) {
     val ctx = object : CompilationContext {
@@ -34,19 +26,9 @@ fun convert(arguments: ConverterArguments) {
         Format.OpenAPIV3 -> OpenAPIV3Parser
         Format.Avro -> AvroParser
     }
-    val options = arguments.parseOptions()
     arguments.input
         .map { ModuleContent(it.name.value, it.content) }
         .map { moduleContent -> parser.parse(moduleContent, arguments.strict) }
-        .map { ast ->
-            ast.copy(
-                modules = ast.modules.map { module ->
-                    module.copy(
-                        statements = validate(options)(module.statements).getOrElse { error("Invalid statements") },
-                    )
-                },
-            )
-        }
         .flatMap { ast ->
             arguments.emitters.flatMap {
                 it.emit(ast, arguments.logger)
